refactored create quiz
