+++ src/components/CreateQuiz/MakeQuiz/MakeQuiz.tsx
@@ -1,36 +1,11 @@
 import { Typography } from "@material-ui/core";
-import {
-  Box,
-  Button,
-  Chip,
-  CircularProgress,
-  FormControl,
-  Grid,
-  IconButton,
-  Input,
-  InputLabel,
-  Paper,
-  Table,
-  TableBody,
-  TableCell,
-  TableContainer,
-  TableHead,
-  TablePagination,
-  TableRow,
-  TableSortLabel,
-  TextField,
-} from "@mui/material";
+import { Box, Button, Chip, Grid, Paper } from "@mui/material";
 import { useState } from "react";
-import { getTotalSubmittedQuizInfo } from "../../api/apiAgent";
-import { submittedQuizResponse } from "../../Interface/QuizDetails";
-import { getComparator } from "../../utils/TableSortFunctions";
-import { columns } from "../SubjectExpertComponents/QuestionSetsTableColumn";
+import { getTotalSubmittedQuizInfo } from "../../../api/apiAgent";
+import { submittedQuizResponse } from "../../../Interface/QuizDetails";
 import "./MakeQuiz.style.scss";
-import PastEvaluationsTable from "./PastEvaluationsTable";
-import SearchIcon from "@mui/icons-material/Search";
-import InputAdornment from "@mui/material/InputAdornment";
-import { VisibilityOff, Visibility } from "@material-ui/icons";
-import SearchQuestionSets from "./SearchQuestionSets";
+import PastEvaluationsTable from "../PastEvaluationsTable/PastEvaluationsTable";
+import SearchQuestionSets from "../SearchQuestionSets/SearchQuestionSets";
 
 const favSub = ["React", "javascript", "c#"];
 
@@ -51,8 +26,6 @@ const MakeQuiz = () => {
         // setLoader(false);
         console.log("error in total quiz info api");
       });
-
-    // get the submitted quiz response for this particular tag and dispaly it in a table
   };
 
   return (
