+++ src/components/CreateQuiz/SelectQuestions/SelectQuestionsModal.style.scss
@@ -1,0 +1,14 @@
+.modal-container {
+  min-height: 70vh;
+}
+
+.close-modal-button {
+  display: flex;
+  justify-content: center;
+  margin-top: 1rem;
+}
+
+.questions {
+  display: flex;
+  align-items: center;
+}
