--- src/components/CreateQuiz/SelectQuestionsModal.style.scss
@@ -1,37 +1,0 @@
-// .preview-modal-container {
-//   border: 2px solid red;
-//   margin-left: 5%;
-//   margin-right: 5%;
-//   margin-top: 15vh;
-//   // background-color: #fffef0;
-//   min-height: 50vh;
-//   max-height: 80vh;
-//   overflow-y: scroll;
-
-//   .questions-container {
-//     //  margin-bottom: 2rem;
-//     display: flex;
-//     justify-content: space-between;
-//     align-items: center;
-//   }
-// }{
-
-.modal-container {
-  // border: 2px solid red;
-  min-height: 70vh;
-
-  .text {
-    // width: 70%;
-  }
-}
-
-.close-modal-button {
-  display: flex;
-  justify-content: center;
-  margin-top: 1rem;
-}
-
-.questions {
-  display: flex;
-  align-items: center;
-}
